+++ src/app.test.js
@@ -28,7 +28,7 @@ describe('Application - JSDOM environment', () => {
         version: 'bCsMYVYVawc8SMPzZWJpiw',
       },
     };
-    const resolvePageAssetCall = () => Promise.resolve(pageData);
+    const resolvePageAssetCall = () => Promise.resolve({ data: pageData, status: 200 });
     const fakeSdk = { assetByVersion: resolvePageAssetCall, assetByAlias: resolvePageAssetCall };
     const store = configureStore({ initialState: {}, sdk: fakeSdk });
     const div = document.createElement('div');